+++ lib/loader.ex
@@ -17,8 +17,10 @@ defmodule Loader do
     %Country{
       name: country["name"],
       full_name: country["full_name"],
+      informal_short_name: country["informal_short_name"],
       alpha_2: country["alpha_2"],
       alpha_3: country["alpha_3"],
+      legacy_alpha_3: country["legacy_alpha_3"],
       numeric: country["numeric"],
       tld: country["tld"]
     }