+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 Initial release
 
 # 0.1.1 (2023-03-11)
+
+- Added functions that raise on no match
+- Added legacy alpha-3 code "ROM" for Romania
+- Added functions to get country by legacy alpha-3 code
+- Added informal short names for UK and USA